requirement.txt
Django==2.2
flake8>=3.8.3,<3.9.0
psycopg2>=2.8.5,<2.9.0
gunicorn>=20.0.4,<20.1.0
Pillow==7.2.0
django-allauth==0.42.0
django-crispy-forms==1.9.1
python-dateutil>=2.8.1,<3.0.0